Remove gettext-related files and include gettext in configure.ac
Since we build for different versions of platforms, some old, some new,
the gettext-related included files were starting to have issues with the
latest versions of autotools used to configure the program.

Remove all gettext-related files and include AM_GNU_GETTEXT in configure,
also requiring on intltool to be installed.